Refactoring of the Token authentication for better testability. Adding unit tests for AuthTokenProvider
